mapp AccountCommentsAndNotes component tests to AC
