+++ cypress/component/manualAccountCreation/FinesMacAccountCommentsAndNotes/constants/fines-mac-account-notes-and-comments-elements.ts
